Fix outdated comments and documentation, and split up a file.
I moved DistributedClosestPointImpl into its own file to avoid
confusing its code with DistributedClosestPoint, which has some
similar methods.
